third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-auto-whitespace-removing-expected.txt

This test verifies that auto-appended spaces are removed on consequent enters.

function (){}
    if (a == b) {


Running: testCollapsedBlock
function.(){

....|
}
....if.(a.==.b).{


Running: testOpenCurlyBrace
function.(){}
....if.(a.==.b).{

........|


Running: testSmartIndent
function.(){}
..

..|..if.(a.==.b).{


Running: testMultiCursorSelection
function.(){}
..

..|..

....|if.(a.==.b).{


Running: testEditedAutoIndent
function.(){}
....if.(a.==.b).{
........W
........|
